[Optimize] Optimize the automatic generation logic for JNI files and SO load files.
Summary:
Previously, the automatic generation of JNI headers and SO load files was implemented, addressing the issue of manual maintenance for SO load files. However, several problems remain, such as the fragmentation of JNI config files, the need for secondary encapsulation when invoking automation scripts, and the absence of support for incremental generation of JNI headers.

Key Modifications:
1. Treat the SO load file as an object, centralizing all configurations in the JNI config file. The JNI generator toolchain now accepts a root path and a YAML file for JNI configs, enabling the generation of corresponding SO load files, GN scripts, and JNI headers based on input parameters without the need for secondary encapsulation.
2. Enable incremental generation of JNI files to reduce time consumption.
3. Optimize certain logic within the JNI automatic generation toolchain.

AutoSubmit:true
SkipChecks:macro