[Optimize] Move PageConfig from NativeFacade to LynxContext.
PageConfig should be a shared object at the TemplateBundle level, and it's not quite suitable to place it in NativeFacade.

For now, it's moved to LynxContext, which reduces the number of member variables in LynxContext itself. Additionally, we consolidate the pageConfig-related switches within LynxContext, making it easier to maintain.

Further optimization can be done later so that the LynxView corresponding to the same TemplateBundle can share a single PageConfig object.

AutoSubmit: true
issue: m-6564832587