[Optimize][DevTool] Lazy load error parser for logbox while parsing error
This is a preparatory MR for migrating LogBox from LynxDevTool to module BaseDevTool. In this MR, we have decoupled the code related to Lynx into a standalone frontend project lynx-error-parser and implemented lazy loading of lynx-error-parser's production to ensure that the code migrated to BaseDevTool remains fully independent of Lynx.