[Optimize] iOS UI node array by not copying and freeing memory after use
Each node in iOS LynxUI may create three NSMutableArrays to store the node's actions. When flushing, the arrays are copied and the original arrays are cleared. However, this does not release the array itself (NSMutableArray struct, 48 bytes) and the array's buffer memory. Now it is modified to set it to nil after use.

issue: m-6383995029