[Optimize] Use relative path in gn files.
In order to build Lynx SDK without the help of lynx's parent folder, we should use relative path in gn files.
For convenience, we have developed a script tool to assist in converting GN relative paths.
The usage of this tool is: execute the command [python3 tools/gn_tools/gn_relative_path_converter.py .] in the lynx directory.
In addition, we add a script to check gn absolute path and automatically convert it to relative path.