[Feature] Using LynxViewGroup to reuse lynx runtime environment
In this commit, we introduce a new Concept: `LynxViewGroup` to manager the reusement of lynx runtime environment. Each lynxView may build with a same LynxViewGroup instance such that LynxViewGroup is shared between multiple lynxViews.

currently, LynxViewGroup supports with below properties:
- `behaviorRegistry`
- `lynxRuntimeOptions`
- `contextData`
- `threadStrategy`
- `enableAutoExpose`
- `enableLayoutSafepoint`
- `density`

and custom properties as:
- `url`
- `templateBundle`
- `globalProps`

In LynxTemplateRender, we use `ILynxViewConfigProvider` to decouple with LynxViewBuilder.

later, we will support using `LynxViewGroup` as `RuntimeCacheManager` that's why a new interface `ILynxViewRuntimeCacheManager` is introduced

issue: f-6578369019
AutoSubmit: true