[BugFix][iOS] ColorLayer not being removed.
Should reset fillColor when backgroundColor is reset to nil and backgroundImage
not changed.

Before this patch, the layers are reset when backgroundImages array changed or colorChanged.
If the backgroundImage is not changed and color changed to nil, the reset process is skiped.

Fixed it by reseting the fill color before applying gradient layers.

issue: f-6733330220