[Optimize][Animation] Fix same target value trigger transition recursively
The same target value should not trigger transition, Align web presentation.
At the beginning of the transition, record its endvalue. If the endvalues are the same, the animation will not be triggered repeatedly. When the transition ends or is destroyed, clear the recorded endvalue

issue: m-0
AutoSubmit: True
AutoLand: release/3.3, release/3.2