[Optimize][DevTool][Part1] Replace devtoolEnabledForDebuggableView by instance-level debuggable bool value
summary: We need to remove this judgment and use the debuggable value.

In order to pass it in from the time of initializing the LynxView to
determine whether to enable the debugging switch at the single LynxView
instance-level. This is the first part of this series of changes, and we
will add it later.

issue: m-6736726249