[BugFix] Making sure that TemplateData is released properly
It's needed to make sure the templateData is released with the engine thread. In this patch
an asscociated lynxContext is bind to templateData we will post the recycle runnable to
engine thread if lynxContext is valid.

thus that we have no need to keep the updateDataList at templateRender as it may cause memory issues
when lynxView is kept alive the whole app lifecycle, the updateDataList will be larger and larger.

anding a trail key "enable_data_list_fix" to control the fix with TrailService

issue: f-6765321515
AutoSubmit: true
AutoLand: release/3.5