[BugFix][Clay] Fix issue that clay's input-view cannot successfully render text on Android.
We found an issue that on Android, Clay's input-view cannot successfully render text and cursor. This is because the transform we set for the drawing is incorrect, and this change will fix it. What we want is only a flip-Y matrix for the original transform.

AutoLand: release/3.5, release/3.6

issue: m-6823587284